requirements.txt
torch
torchvision
onnx
onnxruntime
numpy
opencv-python
tqdm
pyyaml
pandas
scikit-learn
albumentations
mmcv-full
timm
matplotlib
